Implement beliefs.json support
This is implemted by adding BeliefSpec, based on BehaviourSpec. This is
tested in BeliefSpecTest. This works by reading from the JSON into an
array of BeliefSpec. This is then par mapped to BasicBeliefs.

This is added as an option to the CLI -c or --belief (-c is for
concept).

This is documented as an option in docs/concept.1.adoc.

The format of beliefs.json is given in docs/beliefs.json.5.adoc, with an
example being given in beliefs.json.5.example.

Signed-off-by: Robert Greener <me@r0bert.dev>